Move EMA and stream throttling code to a new file
